hoist the state used for backend config into meta
We want to use the same data structure to store the plugin digests as
well as the configured plugin vendor path. Move the State used for
storing the config into the command.Meta struct.

We can remove the state.State arguments from the meta backend methods,
and retrieve the current state using the InitState method. We also
supply a SetInitState method for running from a plan, where the config
state is overridden by the state from the plan.